Get all game data from the API and handle user changes
Move HTTP code from Main to GameForm.  We'll need the form to fetch
the latest game data since it will also be handling end-user changes.

Detect when text inputs receive and lose focus.  In a future commit,
we'll make sure that the text input value isn't updated while the
component is in focus.

When the text input value changes, change the color and stop updating
it from the server.  The value shouldn't change until the user hits
Save or Undo.

Handle Undo.  Remove dummy components.  We'll re-add them when it's
time to implement.