Implement API update
Implement a Tanstack Query mutator.  When the user clicks the
Save button, send a PATCH request to the API and update the
client cache with the Game data that's returned in the API
response body.

Don't destructure the fields returned by `useQuery()` to clearly
separate the query from the mutation.  Remove `console.log()`
calls that were used for debugging.