Fix unit tests
`npm test` was failing on importing **axios** in **api.ts**.  I eventually
discovered that moving the axios code to **ErrorBoundaryFallback.tsx**
caused the test failure to go away.

Luckily `handleAxiosError()` was only called from this component.
We'll have to dig into this if we ever have to write utility functions
that wrap Axios.