Exponential backoff when viewing a game
Implement exponential backoff in `ViewGameComponent` so that we hit
the API less frequently when data is not changing.  Rename `ViewGame`
to `ViewGameComponent` so that the component name doesn't collide with
the model name.  Use the `ViewGame` model class instead of `Game`
since the **/api/games/view_id** endpoint doesn't return all of the
fields on the `Game` object.