Handle burst requests to API
Add `burst` config to API endpoints.  This causes NGINX to handle up
to 10 burst requests to the API.  If the client sends a burst of API
requests, NGINX blocks and handles them at the rate of 2 per second.
When more than 10 requests from a client back up, NGINX returns 503.

This handles the case when the user runs multiple instances of the
app in separate tabs or React/Tanstack Query hits the API more often
than we'd like.