Show modal dialog when copying links to clipboard
When copying the view or admin link to the clipboard, show a modal
dialog instead of an alert.  This places the message at the top of
the screen instead of below the fold.

Consolidate the view and admin alerts into one.  We will probably
remove the admin link from the UI.

fixes #2